Started work on site edit page
Added site-edit.html and the link to it in function site_edit_app().
Added siteDat.json to userData site directories, which will contain the data/structure of the website.
Added .localnav styling in main.css
Added a way to change the colors of the website in site-create-options-1.html. This only works for the primary color and doesnt work out the lighter and darker versions as the rgb to hsl code isnt working for some reason.